// src/interfaces/appInterfaces.tsx
export interface PesadasResponse {
    _id?: string
    fecha: string, 
    hora: string, 
    lote: string, 
    caravana: string, 
    peso: string,
    raza: string,
    tropa: string,
    estado: string,
    genero: string, 
}

export interface VersionResponse {
    _id: string,
    version: string,
}

// Generated by https://quicktype.io

export interface Upgradeversion {
    version:  number;
    upgrades: { [key: string]: Array<string[]> };
}